ODBC: add casts to ODBCLOG calls. All arguments to ODBCLOG (fprintf) that are not of type (unsigned) int but are printed using %d (%u) now have a cast. Also make sure that unsigned values are printed using %u. Finally, it makes no sense using the %hd/%hu format in printf and friends: the argument is converted to (unsigned) int by the compiler before printf sees it.

+++ clients/odbc/driver/SQLBindCol.c
@@ -52,7 +52,9 @@ SQLBindCol(SQLHSTMT hStmt,
 	ODBCDesc *desc;		/* Application Row Descriptor */
 
 #ifdef ODBCDEBUG
-	ODBCLOG("SQLBindCol " PTRFMT " %d %d " LENFMT "\n", PTRFMTCAST hStmt, nCol, nTargetType, LENCAST nTargetValueMax);
+	ODBCLOG("SQLBindCol " PTRFMT " %d %d " LENFMT "\n",
+		PTRFMTCAST hStmt, nCol, (int) nTargetType,
+		LENCAST nTargetValueMax);
 #endif
 
 	if (!isValidStmt(stmt))
